se crean metodos para el controller

+++ src/main/java/com/adoptable_companions/adoptable_companions/repository/AnimalRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.adoptable_companions.adoptable_companions.entity.AnimalEntity;
 
 @Repository
-public interface AnimalRepository extends JpaRepository<AnimalEntity, Integer> {
+public interface AnimalRepository extends JpaRepository<AnimalEntity, Long> {
 }
